Add loading feature
